PythonToolbox/requirements.txt
wheel
setuptools
twine
sqlalchemy
pika
influxdb
python-dotenv
python-keycloak
kafka-python
